backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.0
dj-database-url==2.1.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
tzdata==2025.2
gunicorn==23.0.0
whitenoise==6.9.0
python-dotenv==1.0.1
